requirements.txt
streamlit
python-dotenv
qdrant-client
sentence-transformers
langchain
langchain-community
PyMuPDF
google-generativeai==0.8.5
protobuf==4.25.3
cohere
